+++ src/rules/index.js
@@ -1,5 +1,6 @@
 import atExtendNoMissingPlaceholder from "./at-extend-no-missing-placeholder"
 import atFunctionPattern from "./at-function-pattern"
+import atImportNoPartialExtension from "./at-import-no-partial-extension"
 import atImportNoPartialLeadingUnderscore from "./at-import-no-partial-leading-underscore"
 import atMixinNoArgumentlessCallParentheses from "./at-mixin-no-argumentless-call-parentheses"
 import atMixinPattern from "./at-mixin-pattern"
@@ -10,6 +11,7 @@ import selectorNoRedundantNestingSelector from "./selector-no-redundant-nesting-
 export default {
   "at-extend-no-missing-placeholder": atExtendNoMissingPlaceholder,
   "at-function-pattern": atFunctionPattern,
+  "at-import-no-partial-extension": atImportNoPartialExtension,
   "at-import-no-partial-leading-underscore": atImportNoPartialLeadingUnderscore,
   "at-mixin-no-argumentless-call-parentheses": atMixinNoArgumentlessCallParentheses,
   "at-mixin-pattern": atMixinPattern,